move StyleHandler to a tsx file
